Updated Poco X3 ROM Data (#6)
* updated status of discontinued roms and moved them down, fixed some rom links

* updated status of discontinued roms for poco x3

* Back to the old order

It's a list. The order does not matter. A function to sort the list will be added to the website in the future.

* Remove CorvusOS completely

It has been dead for years now. RIP

* Remove unnecessary space

* PixysOS back to official

This build is not marked as "deprecated" by PixelOS' team.

* CherishOS back to official

This build is not marked as "Discontinued" by the CherishOS team.

* "Inactive" -> "Discontinued"

For consistency.

---------

Co-authored-by: Piotr Lange <79166768+PiotrZPL@users.noreply.github.com>